Verify WhisperAPI version on startup (#32)
* Add parameter for whisper API version

Refactor to allow different versions

* Refactor: Decouple tests from WhisperAPI versions

We want most of the tests to be independant of the WhisperAPI version. Only the
whisper_api_adapter needs to know about all the versions and translate it
to our classes.

* Support Whisper API version 11

The code is now easier to extend for new versions. This was required
because we realized the API contract had changed and might change again in
the future.

* Verify WhisperAPI version on startup

* Remove command line argument for api version

Run the automatic test with the version from WhisperAPI

* Fix division by zero bug when calculating intent accuracy index

* No longer change the suggestion's data case

The reason is that URL's are case sensitive and are not garanteed to work in lowercase

* Support WhisperAPI version 12